+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * `AIXM::PayloadHash` class
 * `mkmid` executable to insert `mid` attributes into valid OFMX file
 * `ckmid` executable to check `mid` attributes in an OFMX file
+* `AIXM::Component::Geometry#point?|circle?|polygon?`
 
 ### Breaking Changes
 * Require Ruby 2.7